refactor:project,projectmember,task-service,serviceImpl,Request,Response

+++ spring-breeze-erp-v3/back/bootSBerp/src/main/java/com/sb/erp/proj/service/ProjectServiceImpl.java
@@ -29,11 +29,11 @@ public class ProjectServiceImpl implements ProjectService{
 	@Override public int insert(ProjRequest dto) {  return dao.insert(dto); }
 
 	//프로젝트 상세보기
-	@Override public ProjResponse select(int proId) {  return dao.select(proId); }
+	@Override public ProjResponse select(Long proId) {  return dao.select(proId); }
 	
 	//프로젝트 삭제
 	@Transactional //세 쿼리 다 성공 → 트랜잭션 커밋 (전부 반영)
-	@Override public int delete(int proId) {  
+	@Override public int delete(Long proId) {  
 		 // 1. 태스크 삭제
 		  dao.deleteTaskByProjectId(proId);
 
@@ -47,7 +47,7 @@ public class ProjectServiceImpl implements ProjectService{
 	@Override public int edit(ProjRequest dto) {  return dao.update(dto); }
 	
 	//프로젝트 수정뷰
-	@Override public ProjResponse editView(int proId) {  return dao.select(proId); }
+	@Override public ProjResponse editView(Long proId) {  return dao.select(proId); }
 
 	/*paging*/
 	@Override public List<ProjResponse> selectAll(ProjectSearchRequest search) {
@@ -58,7 +58,7 @@ public class ProjectServiceImpl implements ProjectService{
     @Override public int selectCnt(ProjectSearchRequest search) { return dao.selectCnt(search); }
 
 	//Ai 결과 보고서
-	@Override public ProjectAnalysisResponse projectAnalysis(Integer proId) {
+	@Override public ProjectAnalysisResponse projectAnalysis(Long proId) {
 		ProjectAnalysisResponse dto = dao.projectAnalysis(proId);
 		if(dto==null) {return null;}
 		long remainDays = ChronoUnit.DAYS.between(LocalDate.now(), dto.getEndDate());
@@ -67,7 +67,7 @@ public class ProjectServiceImpl implements ProjectService{
 		return dto;
 	}
 	//디스코드 알람 전송
-	@Override public String analyzeProject(Integer proId) {
+	@Override public String analyzeProject(Long proId) {
 		ProjectAnalysisResponse dto = projectAnalysis(proId);
 		if(dto==null) {return "프로젝트 정보를 찾을 수 없습니다.";}
 		String result = openAi.analyzeProject(dto);
@@ -75,7 +75,7 @@ public class ProjectServiceImpl implements ProjectService{
 				return result;
 	}
 	//주간 보고서
-	@Override public WeeklyReportResponse weeklyReport(Integer proId) {
+	@Override public WeeklyReportResponse weeklyReport(Long proId) {
 		WeeklyReportResponse dto = dao.weeklyReport(proId);
 	    if (dto == null) { return null; }
 	    long remainDays = ChronoUnit.DAYS.between(LocalDate.now(), dto.getEndDate());
